Move volume cli into component.
