+++ components/volume/client/cmd.go
@@ -1,4 +1,4 @@
-package volume
+package client
 
 import (
 	"fmt"
