+++ components/volume/client/list.go
@@ -1,4 +1,4 @@
-package volume
+package client
 
 import (
 	"sort"
