+++ daemon/volumes_unix.go
@@ -144,7 +144,7 @@ func (daemon *Daemon) verifyVolumesInfo(container *container.Container) error {
 
 			if strings.HasPrefix(hostPath, vfsPath) {
 				id := filepath.Base(hostPath)
-				v, err := daemon.volumes.CreateWithRef(id, volume.DefaultDriverName, container.ID, nil, nil)
+				v, err := daemon.volumeComponent.Create(id, volume.DefaultDriverName, container.ID, nil, nil)
 				if err != nil {
 					return err
 				}
